Removed withStyles HOC
